+++ tests/test_dataset.py
@@ -562,10 +562,13 @@ class Test_TestDataset(CustomTestCase):
         obj_det_ds = ObjectDetectionDataset(
             input_csv_path=csv_path,
             root_dir=os.path.dirname(csv_path),
-            augmentation_list=[A.Normalize(), A.pytorch.ToTensorV2()],
+            augmentation_list=A.Compose(
+                [A.Normalize(), A.pytorch.ToTensorV2()],
+                bbox_params=A.BboxParams(format="coco", label_fields=["labels"]),
+            ),
         )
         self.assertEqual(len(obj_det_ds), 12)
-        image, target, id = obj_det_ds[0]
+        image, target = obj_det_ds[0]
         self.assertEqual(image.shape, (3, 571, 571))
         self.assertEqual(target["boxes"].shape, (2, 4))
         self.assertEqual(target["labels"].shape, (2,))
@@ -575,10 +578,13 @@ class Test_TestDataset(CustomTestCase):
         obj_det_ds = InstanceSegmentationDataset(
             input_csv_path=csv_path,
             root_dir=os.path.dirname(csv_path),
-            augmentation_list=[A.Normalize(), A.pytorch.ToTensorV2()],
+            augmentation_list=A.Compose(
+                [A.Normalize(), A.pytorch.ToTensorV2()],
+                bbox_params=A.BboxParams(format="coco", label_fields=["labels"]),
+            ),
         )
         self.assertEqual(len(obj_det_ds), 12)
-        image, target, id = obj_det_ds[0]
+        image, target = obj_det_ds[0]
         self.assertEqual(image.shape, (3, 571, 571))
         self.assertEqual(target["boxes"].shape, (2, 4))
         self.assertEqual(target["labels"].shape, (2,))